[Deployment Revisited][Staging] Add endpoint placeholder for crash replication in appengine (#4842)
### Motivation

We intend to sample crashes during fuzz task output, and redirect them
to the upload_testcase handler in appengine. For that, we need to
trigger an appengine endpoint with a pubsub push notification.


This PR adds the 'upload_testcase/crash_replication' endpoint, that will
be triggered by a pubsub push notification. At its current state, it
will only log and emit a metric. The actual implementation will be done
in the future


### Testing strategy

Deployed to the development environment, and published a message with
the following content manually: {1:2}

Verified logs were correctly created, and a metric was emited:


![image](https://github.com/user-attachments/assets/49617982-7387-4187-b8a7-5a9c7dc2afe4)

![image](https://github.com/user-attachments/assets/39921d71-fd7e-4722-9905-cb849aeac505)