[Logging] Truncate extras when dumped as a string in labels (#4852)
Background: As commented in other PRs
(https://github.com/google/clusterfuzz/pull/4845,
https://github.com/google/clusterfuzz/pull/4752), we have a recurring
issue with oversized logs.

This PR adds the truncate method in another part of the code, where the
`extras` field from python logging is dumped as a json string in the
labels from GCP logging. I'm not sure if this label is useful and/or
being used anymore, but to avoid removing for now, just inserting the
truncate function after doing the `json.dumps` is enough.

This can be done because the value in labels is not interpreted as a
json by the GCP logging afterward, but as a raw string, as seen here:
https://screenshot.googleplex.com/8RG3K5aryJsjudm.png

b/414397533