fix: implement check and parse for namedtuple on truncate (#4863)
We had an error when trying to truncate a NamedTuple, and the cause is
that we are trying to deal with it as a `tuple` as tuple is a supertype
of namedtuple.

The fix for it is having a checking line that deals with namedtuple
properly. And the way to indentify it is checking if the type is tuple,
and if the object has the attribute `_fields`.

Signed-off-by: Javan Lacerda <javanlacerda@google.com>