[Deployment revisited][Staging] Sample crashes in fuzz task postprocess and reingest through appengine (#4853)
### Motivation

In order to get test coverage for analyze task in our testing
environments (dev and staging), we need to redirect fuzz task crashes
through the upload testcase handler in appengine. This will trigger
analyze task, allowing this to run without people having to manually
input testcases.

This PR implements the sampling logic during fuzz task postprocess, by
extending project.yaml with two new parameters:


```
fuzz_task_sampling:
    sampling_topic: X
    sampling_rate: Y
```

Sampling topic is the topic to which the messages will be published, to
be ingested by appengine. Sampling rate is expected to be an integer
from 0 to 100, representing the probability that a crash group will be
sampled within fuzz task.

### Implementation

The Crash class in fuzz task has all the data required for a testcase to
be created through a manual appengine upload.

The message will have the following parameters:

```
{
        'fuzzed_key': 'some-fuzzed-key',
        'job': 'some-job',
        'fuzzer': 'some-fuzzer',
        'target_name': some-binary',
        'arguments': 'some-arg',
        'application_command_line': 'some-cli-command',
        'gestures':  '[gesture]',
        'http_flag': False,
        'original_task_id': 'some-task-id',
}
```

Important to notice, fuzzed_key is the blob id of the crash contents.
Since both appengine and bots understand these blob IDs, it suffices to
pass the key in the message, simplifying the implementation.

The do_post method in upload_testcase.py was refactored. The core logic
was moved to the 'handle_upload' function, in order to allow it to be
reused for requests originating from both http calls and pubsub
messages.

Also, the auth.get_current_user method was extended in order to attempt
to parse the user email from the Bearer header, since this is the
mechanism by which pubsub messages authenticate. This information is
used to check if the SA used by the pubsub notification is a privileged
user, and only then allows the upload to be performed

### Configuration assumptions

This assumes the appengine default service account is a privileged user
in ClusterFuzz

### Testing strategy

Unit tests were added, and a deployment to dev was performed. The end to
end flow works for both engine and black box Fuzzers.

#### Black box fuzzer evidence

https://screenshot.googleplex.com/AEYSHsZRwwouFoF > Successful sample in
fuzz task
https://screenshot.googleplex.com/AaNfUrenGiyBo2u > Successful appengine
upload
https://screenshot.googleplex.com/MnB5bcTczmN2jdA > Successful analyze
repro

#### Engine fuzzer evidence

https://screenshot.googleplex.com/8pfHE7K6WhAFHJG > Successful sample in
fuzz task
https://screenshot.googleplex.com/6zvki8S5WiRYhbm > Successful appengine
upload
https://screenshot.googleplex.com/7cWjiCtjDFGDkgk > Analyze reproduces