[Tracing] Add task events to trusted tasks (#4864)
### Description
This PR instruments the code to emit `TaskExecutionEvents` during
trusted tasks executions.

Trusted tasks run in a single execution (on the same machine) and don't
depend on the pre->main->post stages, thus we only need to emit events
stating that the task started and, after execution, if the task finished
successfully or with an exception:
* Task started: `Event: {Task Status=Started, Outcome=None}`
* Exception during execution: `Event: {Task Status=Exception,
Outcome=UNHANDLED_EXCEPTION}`
* Task Finished: `Event: {Task Status=Finished, Outcome=None}`

(It was necessary to add a try except in order to emit the event in case
of failure)

b/394059581

### Tests
* Unit tests were added to verify the instrumented codepath and check
correctness of the events being emitted.
* Runtime tests  in dev environment:
<img width="1349" height="144" alt="image"
src="https://github.com/user-attachments/assets/6aa189a0-1be3-416a-8afa-79bd4eb00da0"
/>