Adding try catch for truncate (#4866)
Adding try catch to avoid breaking the actual flow then trying to
truncate the logs.

We aren't able to test it properly as is it difficult to force an
execption inside of the try block. But we managed to test it locally by
removing the checker and parser for Namedtuple, having the output in the
image bellow.

<img width="1244" height="325" alt="Screenshot 2025-07-14 at 14 45 58"
src="https://github.com/user-attachments/assets/589acff6-424b-404b-8b07-e4f4290c09a6"
/>

---------

Signed-off-by: Javan Lacerda <javanlacerda@google.com>