[Tracing] Add issue closing events to cleanup (#4876)
This PR instruments the cleanup code to emit the `IssueClosingEvent`.
These events will be emitted every time an issue linked to a testcase is
marked as closed/verified by ClusterFuzz. There are three occurrences of
this:
* Mark an issue fixed if all of its associated reproducible testcases
are fixed.
* Even if the job sets skip auto close issue, clusterfuzz still marks it
as verified and the event is emitted.
* Mark issue as closed if testcase is unreproducible after a certain
deadline.
* In this case, we are also emitting a `TestcaseRejectionEvent`, since
the testcase is also marked as invalid during the cleanup.
* Mark issue as wontfix if testcase is invalid (testcase.fixed='NA')

**Tests**
* Added unit tests to assert that the events are correctly emitted
during cleanup.
* Since the development/staging environments do not run cleanup/triage
crons, it might be needed to run one time against production to verify
the expected behavior.



b/394060579