Allow issue filing without impact task when regression fails. (#4883)
Regression task will fail in a number of expected scenarios such as when
a bug is temporarily fixed and then broken again in a later revision.
Out-of-memory and timeout are particularly vulnerable to this.

There are a number of other issues why regression task can fail. Rather
than hide bugs where this happens, we should report them.