[Grouper] Allow project level definitions to configure the grouper and add metrics (#4885)
**Description**
This PR patches the grouper code in order to add configurations to allo
project level definitions regarding:
* Whether to mark testcases as closed instead of deleting them when
groups overflow (this is intended to avoid data loss when wrong grouping
happens).
* Max group size.
* Crash comparer similarity threshold.
* Crash comparer number of equal frames threshold.

This also adds two metrics for the grouper:
* Count of testcases deleted/closed due to grouper overflow.
* Distribution of group sizes before shrinkage.


**Tests**
Deployed to development environment adding grouper as a cronjob.
* Execution logs: https://screenshot.googleplex.com/9eTbnuNbnDggxay.png
* Groups size metric:
https://screenshot.googleplex.com/5zsftaw6UGrKoXc.png
* Note: Metrics for number of testcases deleted due to group overflow is
not available as it did not happen in development yet.

b/434148261