[Fix] Return minimized keys when libfuzzer cleanse fails (#4882)
A recent bug [b/428886462](https://b.corp.google.com/issues/428886462)
exposed an issue with LibFuzzer cleanse timing out (only used by
OSS-Fuzz).

When this happens, the `last_minimized_keys` is set to None and
postprocess does not handle this, making the testcase minimization to be
set as pending. This PR tries to fix this by using the minimized keys
without cleanse, if it fails. This seems to be the expected behavior,
since there was already a `if cleansed_testcase_path` to check if
cleanse was successfull.