MiraclePtr-protected crashes are not security issues.
As of M128, if a bug is marked MiraclePtr Status:PROTECTED, it is not
considered a security issue. It should be converted to type:Bug and
assigned to the appropriate engineering team as functional issue.

This patch moves the MiraclePtr detection logic from `issue_filer.py` to
`crash_analyzer.py` and uses it to correctly classify MiraclePtr-protected
crashes as non-security issues.

Fixed:https://github.com/google/clusterfuzz/issues/4903
Fixed:https://crbug.com/40930527