Move label is batch from boolean to str. (#4909)
It updates the label is batch to string because it uses the return of
the function `is_uworker` as response, but this function doesn't return
always boolean. For the case where it's not a batch environment, it
returns None. The label is set as a string None for this case.

That's the reason that some datapoints are missing on the dashboard.
Before running the deployment for appengine, the new label was following
the "default" pattern considering all as a String, but after the
deployment, the descriptor has changed the label to Boolean, what stated
to make the metric with `None` to be ignored.

Signed-off-by: Javan Lacerda <javanlacerda@google.com>