Create cloudbuild.yaml for deployment pipeline (#4908)
fix b/437926994

Implement a pipeline for running deployment through butler deploy
script.

The pipeline steps are:

1. Make git clone on clusterfuzz config for a given revision if the
revision is provided as a `substituition`. Substituition is the GCB way
to provide a kind of env vars.
[documentation](https://cloud.google.com/build/docs/configuring-builds/substitute-variable-values)
If the substituition key is not provided, the latest revision is used. 

2. Install deps and run the butler deploy for a also given revision as
substituition. If the revision is not provided, the lastest is used.
These step expected to receive an _PROJECT substituition that points to
the project that should run the deployment, eg: staging, dev, internal,
external...
The way to assure that the pipeline will run against the right
environment is on the trigger definition, where the substituition can be
hardcoded in the trigger definition. We will have a different trigger
for each environment.

Evidence of a successful run
[here](https://github.com/google/clusterfuzz/runs/47932159422)

---------

Signed-off-by: Javan Lacerda <javanlacerda@google.com>