Install git and clone the repository when revision is present (#4913)
I've added a command to install git and clone the clusterfuzz repository
if the clusterfuzz revision is passed as a substituition.

This is needed because GCB doesn't pass the content of the respository
as a git project, then we're not able to change the version in the build
time.

The workaround for this is to clone the repository in build time and
then checkout in the given revision.

---------

Signed-off-by: Javan Lacerda <javanlacerda@google.com>