[Tracing] Emit grouping events (#4925)
Instrumenting grouper code in order to emit grouping events. These
events are only emitted for the testcases being moved to another group
(or to a newly formed one).

I.e., when a pair of testcases is considered similar:
* If both testcases are not grouped yet, an event is emitted for each
one.
* If only one testcase is already assigned to a group, the event is
emitted for the testcase being added to the group.
* If both testcases are in groups, an event is emitted for the one being
moved to the other group. Then for all the remaining testcases that are
also moved due to the group merge.

There is also the case of a testcase being ungrouped because its group
only contains one testcase. This also triggers an event since it is
moving the testcase to an ungrouped state.


Rational at: b/394051778