[Logging] Add relevant testcase data to the structured logs (#4934)
This PR intends to add relevant testcase information to the structured
logs.


**Motivation:** Testcase grouper is also responsible for deleting
testcases when groups overflow beyond the limit we set. In this case, it
is difficult to verify if the grouper was helpful or harmful, since we
can't see the testcase data anymore. This PR also intends to add
relevant testcase information to the deletion logs.

**Test for this log entry by running a local script with the logs
context** (changed "deleting" for "getting" to avoid hindering future
log searches):
* Prints:
  * https://screenshot.googleplex.com/4LSypS9Ff5RPM8t.png
  * https://screenshot.googleplex.com/3fdgw98RenwcrH6.png
  * https://screenshot.googleplex.com/38zReFzqTNDuPzD.png
* Link to logs: https://cloudlogging.app.goo.gl/AWkHF3GXSypCqQnVA

Motivated by b/434148261