Implements feature flag for update_task (#4939)
It updates the task_loop function to run the update task only if the
function `update_task_enabled` retuns True.

The update_task_enabled implements a logic that gets the VM template
metadata called `update_task_enabled` and returns if the environment is
not a Batch environment and this flag is True. We shouldn't run the
update task for Batch because the task is ephemeral for Batch runs.

The flag `update_task_enabled` will be defined by instance template on
the terraform repository, enabling us to rollout the update_task
deprecation by instance group and gcp project. The implementation of the
function will return False, disabling the update task, only if the
metadata exists and the content is different of `true`.

This change was already tested in the dev environment before adding the
validation for Batch environment, follow the screenshot.
<img width="1340" height="527" alt="image"
src="https://github.com/user-attachments/assets/f22b5919-3cea-47f3-999e-d99d94a8c0bd"
/>

---------

Signed-off-by: Javan Lacerda <javanlacerda@google.com>