checkout on the same ref_name (#4942)
It updates the deployment step that gets the config to perform a
checkout on the same branch that is the target for the Clusterfuzz.

Currently we work with two branches: dev and master.

For the current state, when deploying for dev from Clusterfuzz pipeline,
the config version will always be the master, what can override the
current config that can be tested, that is in the dev branch for the
config repository.

This change enforeces to sync the branches during the deployment, so dev
deployment will use the dev branch for config, and the same for master.

---------

Signed-off-by: Javan Lacerda <javanlacerda@google.com>