Fix crash grouping for "NULL" crash states
For security issue, only the crash_state is compared. This is
problematic when there are no crash_state, only a crash_type.

This change explicitly checks for "NULL" in crash states for security
issues and check the crash_type instead.

A regression test has been added to verify the fix.

Bug: 443261679