Fix(docker): Add missing Pipfile to ubuntu-24.04 build
The ubuntu-24.04 docker build was failing because Pipfile and
Pipfile.lock were not being copied into the build context. This change
adds them to the COPY instruction in the Dockerfile.

This change also fixes a bug in the build.sh script where it was
constructing an incorrect Dockerfile path when a version tag was
provided.