Restore cuttlefish fuzzing by ensuring platform initialization runs (#4956)
Due to skipped call to `update_task.run()` when the update task was not
enabled. The `run_platform_init_scripts()` function, which is critical
for booting up the Cuttlefish AVD device, was being bypassed. This
resulted in fuzzing drops due to the device failing to start. Following
change ensures run_platform_init_scripts()' is called even if update
task in not enabled