[Tracing] Create a more lightweight datastore entity for fuzz task events (#4957)
As described in b/441782202, the `task_execution` event from
fuzzer-based tasks (fuzz task and corpus pruning) is responsible for
significantly increasing the storage of events data in Datastore
(passing our expected cost considerations). This high usage mainly stems
from bulit-in index size.

As fuzzer-based tasks are not directly linked to a testcase lifecycle,
this PR creates a separate more lightweight data model for fuzzer-based
task events. This model has fewer indexed fields and a shorter TTL (set
to 2 months for now), as it will only be used to trace clusterfuzz
execution if needed.

A following PR will change the actual emit to use the new event type for
the fuzz and corpus pruning tasks.

Related to: b/441782202 and b/444509656