Run cleanup before tasks even it's not running update task. (#4963)
fix: https://g-issues.chromium.org/issues/446692241

The update_task has several actions that goes inside its execution, and
one of them is cleaning up the environment for the next task. As the
update task was disabled for our platform, we were not cleaning up and
setting up the environment for the next task.

This PR's refact the code enabling the task_loop to prepare the
environment before each task.

Signed-off-by: Javan Lacerda <javanlacerda@google.com>