[Tracing] Fix events timestamp (#4966)
Force usage of UTC when retrieving the timestamp for events. To avoid
having to change the timestamp property in datastore to use `tzinfo`, we
strip out the the timezone information before converting into the
datastore entity.

Per
[documentation](https://googleapis.dev/python/python-ndb/latest/model.html#google.cloud.ndb.model.DateTimeProperty),
datastore assumes that naive datetimes represent UTC.

This should fix the timestamp differences we having between events
emitted from cronjobs and GCE/Batch.

---------

Co-authored-by: André Nogueira Ribeiro <94117783+decoNR@users.noreply.github.com>