Refactor in butler integration tests subparser (#4968)
This PR refactors the butler command-line script by extracting the logic
for adding the `integration_tests` subparser into a dedicated function
for better readability.

No Functional Changes: This is a pure refactor. The behavior of the
butler script remains unchanged.

#### Known Issue

The temporary server process started by `run_server` is not consistently
terminated after `integration_tests` complete. This issue will be
addressed in a future PR as part of a larger effort to improve our test
suite.