[Tracing] Add the back-end structure for the event history table (#4969)
This PR adds the back-end structure for the event history table. A
subsequent PR will use this structure to display a testcase event
history table on the App Engine testcase detail webpage.

Key changes:
- Adds `get_history()` to retrieve all events for a given testcase. The
returned event information includes a URL that points to the GCP logs
associated with the event task.
- Adds `get_task_log()` which retrieves the associated task logs from
GCP.

Related to b/394060581.