Implement immutable image for the base image (#4973)
Fixes b/449765246

I does use the structured implemented in #4971 for implementing the
immutable images.

Here we create the Dockerfile for the immutable base, using the external
config as configuration.

It also add a `push image` flag for pushing the images to the GAR. It's
needed to have two different triggers, one that check if the image can
be built for the pull requests and another one that builds and push the
images for commits on master.

---------

Signed-off-by: Javan Lacerda <javanlacerda@google.com>