fix: Copy the config properly on Dockerfiles. (#4979)
We noticed the way the config was being copied to the Clusterfuzz were
wrong. The project config should be replacing the directory on
src/appengine/config. But actually it was being copied inside of the
src/appengine/config, having something like:
src/appengine/config/chrome-development.

This change to copy and override the src/appengine/config, having this
config with the content of the projects configs.

---------

Signed-off-by: Javan Lacerda <javanlacerda@google.com>