[Tracing] Add task comments to the testcase status board (#4986)
This PR adds task comments to the `Task Events` section in the testcase
status board, which is located on the testcase detail page in the
AppEngine UI. [Here you can check how it looks on the staging
webpage](https://screenshot.googleplex.com/5tVMygX3yXNAwxH).

Key changes:
- Added `task_comments` to `_format_task_execution_event` to display it
in the testcase status board.

Related to b/448109709.