feat(docker): Add BASE_OS_VERSION to Ubuntu 24.04 image (#4997)
This PR adds the `BASE_OS_VERSION` environment variable to the
`ubuntu-24-04` Docker image.

This is a prerequisite for the OS-based task routing feature. Adding
this in a separate, small PR allows the new image to be built and
available sooner, unblocking the main feature rollout.

Once this is merged, bots running this image will have the necessary
environment variable to identify their OS version and pull tasks from
the correct, filtered Pub/Sub subscription.