[Fix] Fix find testcase when dedup only same target is set (#5021)
As described in b/459451185#comment2, there is a failure state where
minimize keeps failing with an exception when the
`DEDUP_ONLY_SAME_TARGET` env is set for a job.

This is happening due to two codepaths that were fixed here:
* Minimimize calls `data_handler.handle_duplicate_entry()` which uses
the `testcase.fuzzer_name` instead of the
`testcase.actual_fuzzer_name()` for the `fuzz_target`.
* `data_handler.find_testcase()` had a misplaced assert that would fail
if the first engine from the list was not the one used by the testcase.

Also, note that this PR changes the behavior to avoid the assert.
Instead, we log a more descriptive error and return None, since failing
to parse the fuzz_target is an expected issue (could be result of
misconfiguration) and should be handled gracefully.

b/459451185