[CLI] feat: add image selection and fix tests (#5024)
### Context

This PR introduces a way to choose between `internal`, `external` and
`dev` immutable images when choosing the project in the `init` command
of the CLI and update the tests accordingly.

For doing that, the old `DOCKER_IMAGE` constant is now a dictionary that
maps the name of the project to the gcr path of the correspondent image.
It was also added the option `--project` (or `-p`) to specify the
project that the user wants to use and pull the correspondent image in
the initialization. The user can select multiple projects, if he wants,
ex.:

```shell
$ casp init -p internal -p dev
```

### Further work

The gcr path to the images are still hardcoded, this can be improved in
future work.